pipecat_resemble/requirements.txt
websockets==11.0
pydantic>=2.0
loguru
pipecat-ai
pyaudio>=0.2.14
dotenv==0.9.9